requirements.txt
streamlit

fastapi
uvicorn[standard]

openai
openai-agents

python-dotenv
pydantic
pydantic-settings
email-validator

sqlalchemy
psycopg[binary]
asyncpg
alembic

chromadb

tiktoken
numpy
pandas

httpx
requests
python-multipart
aiofiles

loguru
tenacity

pytest
ruff